Update Veriform to 1cc06972
This update to Veriform includes a proper implementation of sequence
hashing, where the previous implementation ignored fields containing
sequences when computing the digest:

https://github.com/iqlusioninc/veriform/pull/145

This involved updating the computed digest for the provision message.
However, hopefully this is the last change which will be necessary
(unless there are bugs or other showstoppers).